+++ src/featherlight.js
@@ -54,31 +54,26 @@
 			return opened;
 		};
 
-	var iframeStructure = {
-		// iframeStructure.css({iframeMinHeight: 44, foo: 0}, 'iframe') => {min-height: 44}
-		css: function(obj) {
-			return parseAttrs(obj, 'iframe');
-		},
-		// NOTE: List of available [iframe attributes](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe).
-		attr: function(obj) {
-			var whiteList = ['allowfullscreen', 'frameborder', 'height', 'longdesc',
-					'marginheight', 'marginwidth', 'name', 'referrerpolicy', 'scrolling',
-					'sandbox', 'src', 'srcdoc', 'width'],
-				attrs = {};
-
-			$.each(whiteList, function(index, item) {
-				var attrValue = parseAttrs(obj, 'iframe')[item];
-
-				if (attrValue) {
-					attrs[item] = attrValue;
-					return attrs[item];
-				}
-			});
-
-			return attrs;
+	// Removes keys of `set` from `obj` and returns the removed key/values.
+	function slice(obj, set) {
+		var r = {};
+		for (var key in obj) {
+			if (key in set) {
+				r[key] = obj[key];
+				delete obj[key];
+			}
 		}
+		return r;
+	}
+
+	// NOTE: List of available [iframe attributes](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe).
+	var iFrameAttributeSet = {
+		allowfullscreen: 1, frameborder: 1, height: 1, longdesc: 1, marginheight: 1, marginwidth: 1,
+		name: 1, referrerpolicy: 1, scrolling: 1, sandbox: 1, src: 1, srcdoc: 1, width: 1
 	};
 
+	// Converts camelCased attributes to dasherized versions for given prefix:
+	//   parseAttrs({hello: 1, hellFrozeOver: 2}, 'hell') => {froze-over: 2}
 	function parseAttrs(obj, prefix) {
 		var attrs = {},
 			regex = new RegExp('^' + prefix + '([A-Z])(.*)');
@@ -405,10 +400,12 @@
 				process: function(url) {
 					var deferred = new $.Deferred();
 					var $content = $('<iframe/>');
+					var css = parseAttrs(this, 'iframe');
+					var attrs = slice(css, iFrameAttributeSet);
 					$content.hide()
 						.attr('src', url)
-						.attr(iframeStructure.attr(this))
-						.css(iframeStructure.css(this))
+						.attr(attrs)
+						.css(css)
 						.on('load', function() { deferred.resolve($content.show()); })
 						// We can't move an <iframe> and avoid reloading it,
 						// so let's put it in place ourselves right now: